refactor: move LambdaRequest struct to types package and update handler references
